requirements.txt
alembic==1.1.0
appdirs==1.4.3
attrs==19.1.0
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
exceptiongroup==1.2.0
Faker==2.0.1
Flask==1.1.1
Flask-Assets==0.12
Flask-Compress==1.4.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RQ==0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
gunicorn==19.9.0
honcho==1.0.1
idna==2.8
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==3.1.2
jsonpickle==1.2
Mako==1.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==19.1
pipupgrade==1.5.1
pluggy==1.3.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pyparsing==2.4.2
pytest==7.4.3
python-dateutil==2.8.0
python-dotenv==0.21.1
python-editor==1.0.4
raygun4py==4.3.0
redis==3.3.8
requests==2.22.0
rq==1.1.0
six==1.12.0
SQLAlchemy==1.3.8
text-unidecode==1.2
tomli==2.0.1
urllib3==1.25.8
webassets==0.12.1
Werkzeug==0.15.5
WTForms==2.2.1
